Rename libs/*.h[*.cc] -> libs/*.hpp[*.cpp]
