+++ libs/crypto/CMakeLists.txt
@@ -1,9 +1,9 @@
 add_library(crypto STATIC
-        base64.h
-        hash.h
-        keypair.h
-        crypto.cc
-        crypto.h
+        base64.hpp
+        hash.hpp
+        keypair.hpp
+        crypto.cpp
+        crypto.hpp
         )
 
 target_link_libraries(crypto PUBLIC
