+++ libs/crypto/crypto.cpp
@@ -15,4 +15,4 @@
  * limitations under the License.
  */
 
-#include "crypto.h"
+#include "crypto.hpp"
