+++ libs/crypto/hash.hpp
