+++ libs/crypto/keypair.hpp
@@ -24,7 +24,7 @@
 #include <nonstd/optional.hpp>
 #include <string>
 #include <vector>
-#include "common.h"
+#include "common.hpp"
 #include "ed25519.h"
